89087630 undo delete enhancement

+++ app/src/main/java/com/worldventures/dreamtrips/core/api/spice/DreamTripsRequest.java
@@ -141,16 +141,37 @@ public abstract class DreamTripsRequest<T> extends RetrofitSpiceRequest<T, Dream
 
     public static class DeleteBucketItem extends DreamTripsRequest<JsonObject> {
         private int id;
+        private long delay;
+        private boolean isCanceled = false;
 
-        public DeleteBucketItem(int id) {
+        public DeleteBucketItem(int id, long delay) {
             super(JsonObject.class);
+            this.delay = delay;
             this.id = id;
+        }
+
+        public boolean isCanceled() {
+            return isCanceled;
+        }
+
+        public void setCanceled(boolean isCanceled) {
+            this.isCanceled = isCanceled;
         }
 
         @Override
         public JsonObject loadDataFromNetwork() {
-            Log.d("TAG_BucketListPM", "Sending delete item event");
-            return getService().deleteItem(id);
+            try {
+                Thread.sleep(delay);
+            } catch (InterruptedException e) {
+                e.printStackTrace();
+            }
+
+            if (isCanceled) {
+                return new JsonObject();
+            } else {
+                Log.d("TAG_BucketListPM", "Sending delete item event");
+                return getService().deleteItem(id);
+            }
         }
     }
